# back-end/swagger.json
{
    "swagger": "3.0",
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "AgileOne CRUD",
        "description": "Recruiting Application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:2000/api/",
            "description": "Dev"
        }
    ],
    "components": {        
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "schemes": ["http","https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/admin/get_leads": {
            "post": {
                "tags": ["Consultants"],
                "description": "Get Offers",
                "operationId": "get_leads",
                "summary": "Get Offers Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "perPage": "10/20/30",
                                "page": "Page Number",
                                "search": "Search Text",
                                "role": "Role ID from get_roles api",
                                "user_id": "User ID",
                                "status": "offers/active/exit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/create_lead": {
            "post": {
                "tags": ["Consultants"],
                "description": "Create Offer Details",
                "operationId": "create_lead",
                "summary": "Create Offer Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "candidate_name": "Candidate Name",
                                "direct_client": "Client ID from get_clients api",
                                "end_client": "Client ID from get_clients api",
                                "contact_number": "Candidate Contact Number",
                                "job_id": "Job ID",
                                "job_title": "Job Title",
                                "visa_status": "1-Yes/0-No",
                                "job_duration": "Job Duration",
                                "bill_rate": "Bill Rate",
                                "pay_rate": "Pay Rate",
                                "margin": "Margin Rate",
                                "tentative_start_date": "Tentative Start Date",
                                "employee_id": "Employee ID",
                                "team_lead": "Team Lead ID",
                                "accounts_manager": "Accounts Manager ID",
                                "user_id": "Logged In User ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/update_lead": {
            "post": {
                "tags": ["Consultants"],
                "description": "Update Offer Details",
                "operationId": "update_lead",
                "summary": "Update Offer Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "candidate_name": "Candidate Name",
                                "direct_client": "Client ID from get_clients api",
                                "end_client": "Client ID from get_clients api",
                                "contact_number": "Candidate Contact Number",
                                "job_id": "Job ID",
                                "job_title": "Job Title",
                                "visa_status": "1-Yes/0-No",
                                "job_duration": "Job Duration",
                                "bill_rate": "Bill Rate",
                                "pay_rate": "Pay Rate",
                                "margin": "Margin Rate",
                                "tentative_start_date": "Tentative Start Date",
                                "employee_id": "Employee ID",
                                "team_lead": "Team Lead ID",
                                "accounts_manager": "Accounts Manager ID",
                                "user_id": "Logged In User ID",
                                "id": "Offer ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/get_singlelead": {
            "post": {
                "tags": ["Consultants"],
                "description": "Get Single Offer Details",
                "operationId": "get_singlelead",
                "summary": "Get Single Offer Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "id": "Offer ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/deleteLead": {
            "delete": {
                "tags": ["Consultants"],
                "description": "Delete Offer",
                "operationId": "deleteLead",
                "summary": "Delete Offer Details from system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "user_id": "Offer ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/update_startdate": {
            "post": {
                "tags": ["Consultants"],
                "description": "Update Exit Status / Update Start Date",
                "operationId": "update_startdate",
                "summary": "Update Exit Status / Update Start Date Details from system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "ref": "exit/:empty",
                                "offer_id": "Offer ID",
                                "start_date": "YYYY-MM-DD/:empty"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
...
}